feat: refactor wolf-actions into modular architecture
Fixes #30

Breaking down monolithic 1,043-line wolf-actions.ts file into 9 focused modules:
- auth.ts: Authentication and session management
- client-control.ts: Client start/stop/restart operations
- client-management.ts: Client creation, editing, deletion
- client-retrieval.ts: Client data fetching and listing
- requests.ts: HTTP request utilities and error handling
- server-status.ts: Wolf server status monitoring
- synchronization.ts: Client state synchronization
- types.ts: TypeScript type definitions
- utils.ts: Shared utility functions
- index.ts: Centralized exports maintaining backward compatibility

Key improvements:
- Maintains full backward compatibility through index.ts re-exports
- Follows established patterns from clients/actions refactoring
- Addresses security and maintainability concerns
- Improves code organization and developer experience
- Enables easier testing and future enhancements

All existing imports continue to work unchanged, ensuring zero breaking changes.
