+++ src/app/actions/wolf/index.ts
@@ -1,0 +1,10 @@
+// Re-export all Wolf actions to maintain backward compatibility
+
+export * from './types';
+export * from './auth';
+export * from './client-retrieval';
+export * from './client-management';
+export * from './client-control';
+export * from './requests';
+export * from './synchronization';
+export * from './server-status';
